refactor: follow Rails conventions all over the place
