--- lib/action_view/component/active_model_conversion_monkey_patch.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module ActiveModel
-  module Conversion
-    def to_component_class
-      "#{self.class.name}Component".safe_constantize
-    end
-  end
-end
