+++ lib/action_view/component/test_case.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require "active_support/test_case"
+
+module ActionView
+  module Component
+    class TestCase < ActiveSupport::TestCase
+      include ActionView::Component::TestHelpers
+    end
+  end
+end
