+++ test/config/application.rb
@@ -5,7 +5,7 @@ require File.expand_path("../boot", __FILE__)
 require "active_model/railtie"
 require "action_controller/railtie"
 require "action_view/railtie"
-require "action_view/component"
+require "action_view/component/railtie"
 require "sprockets/railtie"
 
 require "haml"
